// src/app/enums/urls.enum.ts
export enum Urls {
  cabinet = 'cabinet',
  profile = 'profile',
  authorization = 'authorization',
  confirmation = 'confirmation',
  settings = 'settings',
  successRegistration = 'successRegistration',
  home = 'home',
  dashboard = 'dashboard',
  news = 'news',
  tournament = 'tournament',
  diet = 'diet',
  exercises = 'training',
  dayExercises = 'training/day'
}

export enum ApiUrl {
  user = 'user',
  registration = 'user/registration',
  login = 'user/login',
  getAll = 'user/getAll',
  getMe = 'user/getMe',
  update = 'user/update',
  delete = 'user/delete',
  getCurrent = 'user/getCurrent',
  confirm = 'user/confirmation/?token='

}